Cast to unsigned in RAND_RANGE_BADSCALING
Because of the asserts we can be sure that RANG_RANGE_BADSCALING__selected is
never negative. It might exceed the maximum value of a `zend_long`, though, if
`__max` is the maximum value of a `zend_long` and `__min` is negative, causing
undefined behavior.

Thus this commit will not change behavior for all cases that previously *did
not* invoke UB. Specifically all cases where `__max - __min` is smaller than
ZEND_LONG_MAX will not change.